Resolved most crashes
I was able to resolve most crashes, however it still crashes when updating the app against a previous install.

I attempted to save History objects as a second UserDefaults key, however history objects are not saving. Still need to convert to CoreData.

Added detailed notations for a large section of CalculatorViewController